+++ packages/taraxa-ui/src/components/Card/index.ts
@@ -1,1 +1,3 @@
-export { default } from './Card';
+import Card from './Card';
+
+export default Card;
