+++ packages/taraxa-ui/src/components/Header/index.ts
@@ -1,1 +1,3 @@
-export { default } from './Header';
+import Header from './Header';
+
+export default Header;
