+++ packages/taraxa-ui/src/components/SearchInput/SearchInput.stories.tsx
