+++ packages/taraxa-ui/src/components/SearchInput/SearchInput.styles.ts
