+++ packages/taraxa-ui/src/components/SearchInput/index.ts
