--- packages/taraxa-ui/src/components/v1/Card/index.ts
@@ -1,3 +1,0 @@
-import Card from './Card';
-
-export default Card;
