--- packages/taraxa-ui/src/components/v1/Header/index.ts
@@ -1,3 +1,0 @@
-import Header from './Header';
-
-export default Header;
